requirements.txt

selenium
python-dotenv
requests
beautifulsoup4
webdriver-manager
